% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg5-package.R
\docType{package}
\name{demopkg5-package}
\alias{demopkg5}
\alias{demopkg5-package}
\title{demopkg5: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/demopkg5}
  \item Report bugs at \url{https://github.com/ahasverus/demopkg5/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
